+++ src/co/elastic/NotificationManager.groovy
@@ -39,15 +39,16 @@ This method generates flakey test data from Jenkins test results
 def analyzeFlakey(Map params = [:]) {
     def es = params.containsKey('es') ? params.es : error('analyzeFlakey: es parameter is not valid') 
     def secret = params.containsKey('es_secret') ? params.es_secret : null
-    def jobInfo = params.containsKey('jobInfo') ? params.jobInfo : error('analyzeFlakey: jobInfo parameter is not valid')
+    def flakyReportIdx = params.containsKey('flakyReportIdx') ? params.flakyReportIdx : error('analyzeFlakey: flakyReportIdx parameter is not valid')
     def testsErrors = params.containsKey('testsErrors') ? params.testsErrors : []
+    def flakyThreshold = params.containsKey('flakyThreshold') ? params.flakyThreshold : 0.0
     
-    if (!jobInfo || !jobInfo['fullName']?.trim()) {
-      error "Did not receive jobInfo data" 
+    if (!flakyReportIdx || !flakyReportIdx.trim()) {
+      error "Did not receive flakyReportIdx data" 
     }
 
-    def q = toJSON(["query":["range": ["test_score": ["gt": 0.0]]]])
-    def c = '/' + jobInfo['fullName'] + '/_search'
+    def q = toJSON(["query":["range": ["test_score": ["gt": flakyThreshold]]]])
+    def c = '/' + flakyReportIdx + '/_search'
     def flakeyTestsRaw = sendDataToElasticsearch(es: es, secret: secret, data: q, restCall: c)
     def flakeyTestsParsed = toJSON(flakeyTestsRaw)
 
@@ -64,8 +65,19 @@ def analyzeFlakey(Map params = [:]) {
     
     if (ret) {
       githubPrComment(message: msg, commentFile: 'flakey.id')
-    } 
-    
+    }
+}
+
+/**
+This method generates a custom PR comment with the given data
+ * @param file
+ * @param commentFile
+*/
+def customPRComment(Map args = [:]) {
+    def file = args.containsKey('file') ? args.file : error('customPRComment: file parameter is not valid')
+    def commentFile = args.containsKey('commentFile') ? args.commentFile : error('customPRComment: commentFile parameter is not valid')
+    def msg = readFile(file: file)
+    githubPrComment(message: msg, commentFile: "${commentFile}")
 }
 
 /**