+++ src/test/groovy/LicenseScanTest.groovy
@@ -26,12 +26,145 @@ class LicenseScanTest extends ApmBasePipelineTest {
   @Before
   void setUp() throws Exception {
     super.setUp()
+
+    helper.registerAllowedMethod('sh', [Map.class],{m -> return 0})
   }
 
   @Test
-  void test() throws Exception {
+  void testGo() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/*.go'
+    })
     def script = loadScript(scriptName)
     script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa --type go analyze'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'golang'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testNode() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/package.json'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa --type nodejs analyze'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'node:lts'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testRuby() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/*.rb' || m.glob ==  '**/Gemfile'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa --type gem analyze'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'ruby:2.5'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testPython() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/*.py' || m.glob ==  '**/requirements.txt'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa analyze'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testPhp() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/*.php' || m.glob ==  '**/composer.json'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa --type composer analyze'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'composer:1.10.7'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testAnt() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/build.xml'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa --type ant analyze'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'apache-ant'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testMaven() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/pom.xml'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa --type mvn analyze'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'maven'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testGradle() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/build.gradle'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
+    printCallStack()
+    assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
+    assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'fossa --type gradle analyze'))
+    assertTrue(assertMethodCallContainsPattern('sh', 'gradle'))
+    assertJobStatusSuccess()
+  }
+
+  @Test
+  void testNet() throws Exception {
+    helper.registerAllowedMethod('findFiles', [Map.class], {
+      m -> return m.glob == '**/*.csproj'
+    })
+    def script = loadScript(scriptName)
+    script.call()
+
     printCallStack()
     assertTrue(assertMethodCallContainsPattern('getVaultSecret', 'secret=secret/jenkins-ci/fossa/api-token'))
     assertTrue(assertMethodCallContainsPattern('withEnvMask', 'FOSSA_API_KEY'))