+++ test-infra/beats-ci/test_installed_tools.py
@@ -2,15 +2,21 @@ import testinfra
 import pytest
 
 def test_docker_installed(host):
-  cmd = host.run("docker --version")
-  assert cmd.rc == 0, "it is required for all the Beats projects"
+  if host.check_output("uname -m") == "x86_64" :
+    cmd = host.run("docker --version")
+    assert cmd.rc == 0, "it is required for all the Beats projects"
+  else:
+    pytest.skip("unsupported configuration")
 
 def test_docker_compose_installed(host):
-  cmd = host.run("docker-compose --version")
-  assert cmd.rc == 0, "it is required for all the Beats projects"
+  if host.check_output("uname -m") == "x86_64" :
+    cmd = host.run("docker-compose --version")
+    assert cmd.rc == 0, "it is required for all the Beats projects"
+  else:
+    pytest.skip("unsupported configuration")
 
 def test_docker_experimental_configured(host):
-  if host.system_info.type == 'darwin' :
+  if host.system_info.type == 'darwin' or host.check_output("uname -m") != "x86_64" :
     pytest.skip("unsupported configuration")
   else:
     # HOME should point to the USER home for this validation